chore: separated client utils into a separate interface
