+++ src/components/pages/FoundationHeader.astro
@@ -140,7 +140,7 @@ const { mainMenu, ctaButton } = foundationNavigation as {
 </header>
 
 <script>
-  import { stripTrailingSlash } from '@/utils/url'
+  import { stripTrailingSlash } from '@/utils/client'
 
   const siteLinksWrapper = document.querySelector('[data-nav-wrapper]')
   const wideNavMinWidth = window.matchMedia('(min-width: 1060px)')
