+++ src/components/pages/SummitHeader.astro
@@ -180,7 +180,7 @@ const ctaLinkClass = rightNavBase
 </header>
 
 <script>
-  import { stripTrailingSlash } from '@/utils/url'
+  import { stripTrailingSlash } from '@/utils/client'
 
   const summitLinksWrapper = document.querySelector('[data-nav-wrapper]')
   const wideNavMinWidth = window.matchMedia('(min-width: 1060px)')
